requirements.txt
tk
pillow
numpy
gym
torch
